Full Release
Added matrices and "transpose" and "mul" functions. The code-gen is not optimal and is going to be improved - but at least everything works.

Resolved issue with global name collision of "DevTools" - renamed DevTools to "C Sharp Dev Tools". This will generate correct dependencies.

This version is the last version which is not using Unity.Burst 1.5; Burst 1.5 adds BMI intrinsics and more compiler hints. On the managed C# side, more compiler hints are from now on (not this version) going to be used, using C# 8.